requirements.txt
mediapipe
streamlit
cvzone
google-generativeai